Extracting a VideoPageFragment class that displays only video-related information
