+++ app/build.gradle
@@ -65,11 +65,11 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:1.0.0-beta01"
 
-    implementation "androidx.gridlayout:gridlayout:1.0.0"
-
     implementation "androidx.palette:palette:1.0.0"
 
     implementation "androidx.savedstate:savedstate:1.0.0"
+
+    implementation "androidx.fragment:fragment-ktx:1.1.0"
 
     implementation "com.github.hadilq.liveevent:liveevent:1.0.1"
