+++ app/src/main/java/com/javernaut/whatthecodec/presentation/root/viewmodel/ConfigProvider.kt
@@ -1,4 +1,4 @@
-package com.javernaut.whatthecodec.presentation.viewmodel
+package com.javernaut.whatthecodec.presentation.root.viewmodel
 
 import com.javernaut.whatthecodec.domain.VideoFileConfig
