+++ app/src/main/java/com/javernaut/whatthecodec/presentation/video/ui/view/FrameDisplayingView.kt
@@ -1,4 +1,4 @@
-package com.javernaut.whatthecodec.presentation.ui.view
+package com.javernaut.whatthecodec.presentation.video.ui.view
 
 import android.content.Context
 import android.graphics.Bitmap
