+++ app/src/main/res/layout-land/fragment_video_page.xml
@@ -1,47 +1,54 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <FrameLayout
-        android:id="@+id/frameBackground"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <com.javernaut.whatthecodec.presentation.ui.view.FrameDisplayingView
-            android:id="@+id/frameDisplayingView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal" />
-
-    </FrameLayout>
+    android:layout_height="wrap_content">
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="horizontal">
+        android:orientation="vertical">
+
+
+        <FrameLayout
+            android:id="@+id/frameBackground"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+
+            <com.javernaut.whatthecodec.presentation.video.ui.view.FrameDisplayingView
+                android:id="@+id/frameDisplayingView"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_horizontal" />
+
+        </FrameLayout>
 
         <LinearLayout
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_weight="3"
-            android:orientation="vertical">
+            android:orientation="horizontal">
 
-            <include layout="@layout/inline_video_left_panel" />
+            <LinearLayout
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="3"
+                android:orientation="vertical">
 
-        </LinearLayout>
+                <include layout="@layout/inline_video_left_panel" />
 
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="2"
-            android:orientation="vertical">
+            </LinearLayout>
 
-            <include layout="@layout/inline_video_right_panel" />
+            <LinearLayout
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="2"
+                android:orientation="vertical">
+
+                <include layout="@layout/inline_video_right_panel" />
+
+            </LinearLayout>
 
         </LinearLayout>
 
     </LinearLayout>
 
-</LinearLayout>
+</ScrollView>
